requirements.txt
grpcio==1.40.0
grpcio-tools==1.40.0
Autologging==1.3.2
requests==2.27.1
argparse
